move multi-item scoring functions in tokenizer manager into a separate file (#14740)
